Moved framework/oidc to ui-test-app
