+++ test-apps/ui-test-app/src/frontend/appui/backstage/AppBackstageComposer.tsx
@@ -7,7 +7,8 @@ import { connect } from "react-redux";
 import { UserInfo } from "@bentley/itwin-client";
 import { IModelApp } from "@bentley/imodeljs-frontend";
 import { BackstageItemUtilities, BadgeType, ConditionalBooleanValue } from "@bentley/ui-abstract";
-import { BackstageComposer, FrontstageManager, SettingsModalFrontstage, UiFramework, UserProfileBackstageItem } from "@bentley/ui-framework";
+import { BackstageComposer, FrontstageManager, SettingsModalFrontstage, UiFramework } from "@bentley/ui-framework";
+import { UserProfileBackstageItem } from "./UserProfile";
 import { ComponentExamplesModalFrontstage } from "../frontstages/component-examples/ComponentExamples";
 import { LocalFileOpenFrontstage } from "../frontstages/LocalFileStage";
 import { RootState, SampleAppIModelApp, SampleAppUiActionId } from "../..";
